+++ examples/infinite_scroll/android/app/src/main/kotlin/com/example/infinite_scroll/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.infinite_scroll
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()